requirements.txt
llama-stack-client==0.0.57
termcolor==2.5.0
python-dotenv==1.0.1

-e ./agentops